BREAKING CHANGE Add an opional timeout parameter to access token requests.
Previously access token requests did not timeout at all and could result
in the client hanging.  This adds a fairly forgiving request timeout for
the time it takes to connect and get an access token from the auth
service.  The default is overridable from the various entrypoints that
this library provides.

This does not set a default timeout for requests to made using the
clients provided here. See README.rst for more details.